feat: 찜한 경로 조회 구현

+++ src/main/java/com/example/letstrip/repository/TourSpotRepository.java
@@ -7,5 +7,5 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface TourSpotRepository extends JpaRepository<TourSpot, String> {
     TourSpot findByName(String name);
-    TourSpot findById(int id);
+    TourSpot findById(Long id);
 }
